Added privilege check for ADMIN for delete user

+++ client/src/pages/admin/users/UserModal.tsx
@@ -12,6 +12,7 @@ import HistoryIcon from "@mui/icons-material/History";
 import PrivilegeControl from "./PrivilegeControl";
 import { useNavigate } from "react-router-dom";
 import HoldCard from "./HoldCard";
+import Privilege from "../../../types/Privilege";
 
 const StyledInfo = styled.div`
   margin-top: 16px;
@@ -112,15 +113,21 @@ export default function UserModal({ selectedUserID, onClose }: UserModalProps) {
   const handleDeleteUserClicked = () => {
       let fName = getUserResult.data.user.firstName;
       let lName = getUserResult.data.user.lastName;
-      let result = window.confirm(
-          `Are you sure you wish to delete ${fName} ${lName}'s account? This cannot be undone.`
-      );
-      if (result) {
-          deleteUser ({
-              variables: { userID: getUserResult.data.user.id},
-              refetchQueries: [{ query: GET_USER, variables: { id: selectedUserID } }],
-          });
-      };
+      let privilege = getUserResult.data.user.privilege;
+      if (privilege == Privilege.ADMIN) {
+          let result = window.confirm(
+              `Are you sure you wish to delete ${fName} ${lName}'s account? This cannot be undone.`
+          );
+          if (result) {
+              deleteUser({
+                  variables: {userID: getUserResult.data.user.id},
+                  refetchQueries: [{query: GET_USER, variables: {id: selectedUserID}}],
+              });
+          };
+      }
+      else {
+          window.alert("You don't have the privilege level to delete users!")
+      }
   }
 
   return (
